Convert to oci plugin
Instead of attaching to the action / the GitHub Release, we now upload the plugin to our OCI registry.
The upload (push) to the registry is performed by the build tooling (mvn, pnpm).
To this end we also updated the coremedia-plugins-maven-plugin which now requires Java 17.
The studio-client part now uses @coremedia/plugin-manager which requires node 22 and allowance for script execution (to install oras tooling) with onlyBuiltDependencies.

The plugin descriptor is now represented as an OCI manifest and is also pushed via mvn (See configuration in the root pom).

The structure in the registry will be:
* /plugins/content-hub-adapter-filesystem for the descriptor manifest
* /plugins/content-hub-adapter-filesystem/studio-server.content-hub-adapter-filesystem for the studio-server plugin
* /plugins/content-hub-adapter-filesystem/studio-client.main.content-hub-adapter-filesystem for the studio-client plugin

Releases are pushed to repository.coremedia.com/ephemeral-images/ while developer snapshots are pushed to repository.coremedia.com/ephemeral-images/